+++ packages/web/src/components/password-reset/PasswordResetModal.tsx
@@ -5,7 +5,6 @@ import { Modal } from '@audius/stems'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
-import { RESET_REQUIRED_KEY } from 'common/store/account/mobileSagas'
 import EnterPassword from 'components/sign-on/EnterPassword'
 import { AppState } from 'store/types'
 import { isMobile } from 'utils/clientUtil'
@@ -14,6 +13,8 @@ import styles from './PasswordResetModal.module.css'
 import { changePassword } from './store/actions'
 import { getStatus } from './store/selectors'
 const { getNeedsAccountRecovery } = accountSelectors
+
+const RESET_REQUIRED_KEY = 'password-reset-required'
 
 const messages = {
   title: 'Reset Your Password',
